+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/features/user-password.ts
@@ -1,8 +1,8 @@
 import { fillLoginForm, LoginUser } from '../../../auth-forms';
 import { waitForPage } from '../../../checkout-flow';
 import { INPUT_TYPE, MyCompanyConfig } from '../models';
-import { completeForm, FormType } from './form';
 import { loginAsMyCompanyAdmin } from '../my-company.utils';
+import { completeForm, FormType } from './utils/form';
 
 export function userPasswordTest(config: MyCompanyConfig): void {
   const TEST_PASSWORD = 'tE5tP@$5';
