+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/models/my-company.config.ts
@@ -107,11 +107,6 @@ export interface MyCompanyConfig {
   entityIdField?: string;
 
   /**
-   * Test list utilizing a nested tree ux.
-   */
-  nestedTableRows?: boolean;
-
-  /**
    * Configuration of preserve cookies value.
    */
   preserveCookies?: boolean;
@@ -146,11 +141,6 @@ export interface MyCompanyConfig {
    * SUBCATEGORY: Config for modifying user roles.
    */
   rolesConfig?: MyCompanyConfig;
-
-  /**
-   * Set to true to check disabling and enabling an entity
-   */
-  canDisable?: boolean;
 
   /**
    * Set to true to check status in details pane.
