+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/models/my-company.model.ts
@@ -34,7 +34,10 @@ export interface TestListOptions {
 
 export enum MY_COMPANY_FEATURE {
   LIST = 'list',
-  FORM = 'form',
+  NESTED_LIST = 'nestedList',
+  CREATE = 'create',
+  UPDATE = 'update',
+  DISABLE = 'disable',
   ASSIGNMENTS = 'assignments',
   USER_PASSWORD = 'userPassword',
 }
