+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/my-company-features.ts
@@ -1,12 +1,18 @@
 import { assignmentsTest } from './features/assignments';
-import { formTest } from './features/form';
+import { createTest } from './features/create';
+import { disableTest } from './features/disable';
 import { listTest } from './features/list';
+import { nestedListTest } from './features/nested-list';
+import { updateTest } from './features/update';
 import { userPasswordTest } from './features/user-password';
 import { MyCompanyConfig, MY_COMPANY_FEATURE } from './models';
 
 const testMapping = {
   [MY_COMPANY_FEATURE.LIST]: listTest,
-  [MY_COMPANY_FEATURE.FORM]: formTest,
+  [MY_COMPANY_FEATURE.NESTED_LIST]: nestedListTest,
+  [MY_COMPANY_FEATURE.CREATE]: createTest,
+  [MY_COMPANY_FEATURE.UPDATE]: updateTest,
+  [MY_COMPANY_FEATURE.DISABLE]: disableTest,
   [MY_COMPANY_FEATURE.ASSIGNMENTS]: assignmentsTest,
   [MY_COMPANY_FEATURE.USER_PASSWORD]: userPasswordTest,
 };
