+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/my-company.utils.ts
@@ -44,29 +44,6 @@ export function loginAsMyCompanyAdmin(): void {
   cy.requireLoggedIn(myCompanyAdminUser);
 }
 
-export function scanTablePagesForText(
-  text: string,
-  config: MyCompanyConfig
-): void {
-  cy.get('cx-table').then(($table) => {
-    // For table in tree mode expand all elements first and find editable one.
-    if (config.nestedTableRows) {
-      cy.get('cx-org-list div.header button').contains('Expand all').click();
-    }
-
-    if ($table.text().indexOf(text) === -1) {
-      cy.server();
-      cy.route('GET', `**/${config.apiEndpoint}**`).as('getData');
-      // Do not use pagination check for tables in tree mode.
-      if (!config.nestedTableRows) {
-        nextPage();
-        cy.wait('@getData');
-        scanTablePagesForText(text, config);
-      }
-    }
-  });
-}
-
 /**
  * Converts string value to RegExp ignoring case sensivity.
  */
